+++ urisys-node/CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3] - 2026-06-16
+
+### Docs
+- Update README.md
+- Update docs/FLOWS.md
+
+### Test
+- Update tests/test_run_expectations.py
+
+### Other
+- Update scripts/run-lab-unit-ci.sh
+- Update scripts/run_test_sessions.py
+- Update urirdp-docker/packages/python/urirdp/handlers.py
+- Update urirdp-docker/packages/python/urirdp/routes.py
+- Update urirdp-docker/packages/python/urirdp_browser/handlers.py
+- Update urisys-automation-lab/data/events.jsonl
+- Update urisys-automation-lab/flows/03_open_browser_gui.uri.flow.yaml
+- Update urisys-automation-lab/flows/04_browser_download_file.uri.flow.yaml
+- Update urisys-automation-lab/flows/05_fill_form_gui.uri.flow.yaml
+- Update urisys-automation-lab/flows/09_webrtc_video_chat_rdp.uri.flow.yaml
+- ... and 2 more files
+
 ## [0.1.2] - 2026-06-16
 
 ### Docs